rag_service/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
requests
python-dotenv
pydantic
